Account page, clan page, trpc prefetching (#97)
* account page

* lint

* updates

* re-organize

* clean

* cleaned up bad code

* updates to src ssr

* fixed caching

* update lb

* improved code

* dry code

* updating in progress

* finished removing rr stuff

* fix lint

* ifx background colors, oops

* minor fixes

* disable prefetch

* split transitory out

* fix inspection screen mobile

* prefetch trpc queries

* fix some linting

* lite

* update prisma/vercel

* update clan page

* add superjson

* replace urls with fns

* remove progressions for now

* improve clan css

* lint

* gh

+++ src/server/trpc/procedures/user/removeProvider.ts
@@ -1,12 +1,13 @@
 import { z } from "zod"
 import { TRPCError } from "@trpc/server"
 import { protectedProcedure } from "../../middleware"
+import { providerIdToUsernamePropMap } from "~/server/next-auth/providerIdMap"
 
 // delete account from user
 export const removeProvider = protectedProcedure
     .input(
         z.object({
-            providerId: z.string()
+            providerId: z.enum(["discord", "twitch", "twitter"])
         })
     )
     .mutation(async ({ input, ctx }) => {
@@ -14,14 +15,34 @@ export const removeProvider = protectedProcedure
         const providerId = input.providerId
 
         try {
-            await ctx.prisma.account.delete({
-                where: {
-                    provider_userId: {
-                        provider: providerId,
-                        userId: userId
+            const { accounts } = await ctx.prisma.user.update({
+                where: { id: userId },
+                data: {
+                    profile: {
+                        update: {
+                            [providerIdToUsernamePropMap[providerId]]: null
+                        }
+                    }
+                },
+                select: {
+                    accounts: {
+                        where: {
+                            userId: userId,
+                            provider: providerId
+                        }
                     }
                 }
             })
+            if (accounts.length) {
+                await ctx.prisma.account.delete({
+                    where: {
+                        provider_userId: {
+                            provider: providerId,
+                            userId: userId
+                        }
+                    }
+                })
+            }
         } catch (e: any) {
             throw new TRPCError({
                 code: "INTERNAL_SERVER_ERROR",
